requirements.txt
# Core LangGraph workflow
langgraph==0.2.70

# LangChain base (required by langgraph and langchain-anthropic)
langchain==0.3.19

# Anthropic Claude integration — ChatAnthropic class
langchain-anthropic==0.3.10

# Environment variable loading — load_dotenv()
python-dotenv==1.0.1